Lab5 Ex6 + Fix up Ex4 file_block_walk
Regarding fix for `file_block_walk`:

If `f->f_indirect` was already allocated, we would skip retrieving the disk address for that block, resulting in a null pointer check, causing the fs to panic. Solution here would be to ensure that the disk address regardless if the indirect block is allocated or not.