Fix tx_issue for small amounts of packets
There is an issue where at small amount of packets, we would transmit wrong packets from the second packet onwards.

It turns out that when selecting the buffer we want to transmit, we send the wrong index, hence we get a lot of junk data on the line, and is not parsable by tcpdump. Issue is fixed by accessing the right index.

However, this is fixed for small number of packets transmitted. For 100+ packets, we see that there is an issue of sending junk data on the tx line. Still debugging that, I'm using this to save progress.

Also uncommented a line in env to debug.